+++ src/backend/core/admin.py
@@ -194,7 +194,6 @@ class ContactAdmin(admin.ModelAdmin):
     list_display = (
         "full_name",
         "owner",
-        "base",
     )