+++ src/backend/core/api/client/serializers.py
@@ -14,18 +14,12 @@ class ContactSerializer(serializers.ModelSerializer):
         model = models.Contact
         fields = [
             "id",
-            "base",
             "data",
             "full_name",
             "owner",
             "short_name",
         ]
         read_only_fields = ["id", "owner"]
-
-    def update(self, instance, validated_data):
-        """Make "base" field readonly but only for update/patch."""
-        validated_data.pop("base", None)
-        return super().update(instance, validated_data)
 
 
 class DynamicFieldsModelSerializer(serializers.ModelSerializer):